requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
password-strength==0.0.3.post2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
